+++ crates/zng-var/src/lib.rs
@@ -39,6 +39,9 @@ pub use vars::*;
 mod vec;
 pub use vec::*;
 
+mod shorthand_unit;
+pub use shorthand_unit::*;
+
 mod impls;
 
 pub(crate) mod future;
@@ -53,6 +56,9 @@ pub(crate) mod future;
 /// The `U: From<T>` implement is optional, you can use the syntax `fn from(t: T) -> U;` to only generate
 /// the `T: IntoVar<U>` and `T: IntoValue<U>` implementations using an already implemented `U: From<T>`.
 ///
+/// The macro also matches input `_: ShorthandUnit![$ident]` to implement conversions from [`ShorthandUnit!`]. This is
+/// an exact match, you don't need to import [`ShorthandUnit!`].
+///
 /// # Examples
 ///
 /// The example declares an `enum` that represents the values possible in a property `foo` and
@@ -64,23 +70,26 @@ pub(crate) mod future;
 /// pub enum FooValue {
 ///     On,
 ///     Off,
-///     NotSet,
 /// }
 ///
 /// impl_from_and_into_var! {
+///     // generates `From<bool> for FooValue` and `IntoVar<FooValue> for bool`.
 ///     fn from(b: bool) -> FooValue {
 ///         if b { FooValue::On } else { FooValue::Off }
 ///     }
 ///
-///     fn from(s: &str) -> FooValue {
-///         match s {
-///             "on" => FooValue::On,
-///             "off" => FooValue::Off,
-///             _ => FooValue::NotSet,
-///         }
+///     // generates only `IntoVar<FooValue> for Foo`.
+///     fn from(_: Foo) -> FooValue;
+///
+///     // Bind shorthand `On!` with variant
+///     fn from(_: ShorthandUnit![On]) -> FooValue {
+///         FooValue::On
 ///     }
 ///
-///     fn from(f: Foo) -> FooValue;
+///     // Bind shorthand `Off!` with variant
+///     fn from(_: ShorthandUnit![Off]) -> FooValue {
+///         FooValue::Off
+///     }
 /// }
 ///
 /// impl From<Foo> for FooValue {
@@ -91,7 +100,8 @@ pub(crate) mod future;
 /// # pub struct Foo;
 /// # fn assert(_: impl IntoVar<FooValue> + IntoValue<FooValue>) { }
 /// # assert(true);
-/// # assert("on");
+/// # assert(Foo);
+/// # assert(ShorthandUnit![On]);
 /// ```
 #[macro_export]
 macro_rules! impl_from_and_into_var {
@@ -195,11 +205,41 @@ macro_rules! __impl_from_and_into_var {
         }
         //}
     };
+    // INPUT SHORTHAND:
+    (
+        =input=>
+        [$($config:tt)*]
+        (_ : ShorthandUnit![$ident:ident]) $($rest:tt)+
+    ) => {
+        $crate::__impl_from_and_into_var! {
+            =output=>
+            [
+                input_type { $crate::ShorthandUnit![$ident] }
+                $($config)*
+            ]
+            $($rest)+
+        }
+    };
     // INPUT SIMPLE:
     (
         =input=>
         [$($config:tt)*]
         ($ident:ident : $Input:ty $(,)?) $($rest:tt)+
+    ) => {
+        $crate::__impl_from_and_into_var! {
+            =output=>
+            [
+                input_type { $Input }
+                $($config)*
+            ]
+            $($rest)+
+        }
+    };
+    // INPUT SIMPLE DISCARD:
+    (
+        =input=>
+        [$($config:tt)*]
+        (_ : $Input:ty $(,)?) $($rest:tt)+
     ) => {
         $crate::__impl_from_and_into_var! {
             =output=>